requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
python-dotenv
sentence-transformers
langchain-huggingface